refactor(delisting): move Manage liquidity link into toast body (blank-line separated); revert EventToast change (FEDEV-4022)
